core/resnetse/requirements.txt
PyYAML==6.0
webrtcvad==2.0.10
# soundfile==0.12.1
scipy==1.10.1
numba==0.56.4
numpy==1.23.5
# torch==1.13.1+cu116
# torchaudio==0.13.1+cu116
